+++ frontend/lib/widgets/kamar/kamar_card.dart
@@ -12,7 +12,7 @@ class RoomListCard extends StatelessWidget {
     required this.onSelectRoom,
   });
 
-  String _formatPrice(int price) {
+  String _formatPrice(double price) {
     final formatted = price.toString().replaceAllMapped(
       RegExp(r'(\d{1,3})(?=(\d{3})+(?!\d))'),
       (Match m) => '${m[1]}.',
